# `parseModel()` performance improvements