--- a/parse.go
+++ b/parse.go
@@ -56,7 +56,7 @@ type TreeParam struct {
 // Tree is one tree in an XGBoost model. It's the representation we process
 // XGBTree into.
 type Tree struct {
-	Nodes    []*Node // Index 0 is the root.
+	Nodes    []Node // Index 0 is the root.
 	NumNodes int
 }
 
@@ -136,11 +136,7 @@ func parseTree(
 		return nil, fmt.Errorf("getting num nodes as int64: %w", err)
 	}
 
-	var nodes []*Node
-	for i := 0; i < int(numNodes); i++ {
-		nodes = append(nodes, &Node{})
-	}
-
+	nodes := make([]Node, numNodes)
 	for i := 0; i < int(numNodes); i++ {
 		nodes[i].Data = NodeData{
 			BaseWeight:     xt.BaseWeights[i],
@@ -158,8 +154,8 @@ func parseTree(
 			continue
 		}
 
-		nodes[i].Left = nodes[left]
-		nodes[i].Right = nodes[right]
+		nodes[i].Left = &nodes[left]
+		nodes[i].Right = &nodes[right]
 	}
 
 	return &Tree{