Quick perf wins in `parseModel()`: 29.79% less allocs, 6.28% less memory used, 1.41% speed up in runtime.

```sh
$ benchstat old.txt new.txt                        
goos: darwin
goarch: arm64
pkg: github.com/maxmind/xgbshap
              │   old.txt   │              new.txt               │
              │   sec/op    │   sec/op     vs base               │
ParseModel-10   684.4µ ± 1%   674.7µ ± 2%  -1.41% (p=0.011 n=10)

              │   old.txt    │               new.txt               │
              │     B/op     │     B/op      vs base               │
ParseModel-10   186.6Ki ± 0%   174.9Ki ± 0%  -6.28% (p=0.000 n=10)

              │   old.txt   │               new.txt               │
              │  allocs/op  │  allocs/op   vs base                │
ParseModel-10   1.588k ± 0%   1.115k ± 0%  -29.79% (p=0.000 n=10)
```

<details>
<summary>before</summary>

```sh
$ go test -run=^$ -bench=. -benchmem -count=10 | tee old.txt
goos: darwin
goarch: arm64
pkg: github.com/maxmind/xgbshap
BenchmarkParseModel-10    	    1752	    683160 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1737	    683685 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1744	    685918 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1744	    684606 ns/op	  191081 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1748	    683018 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1747	    684094 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1746	    688603 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1759	    799020 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1736	    682506 ns/op	  191080 B/op	    1588 allocs/op
BenchmarkParseModel-10    	    1748	    686241 ns/op	  191080 B/op	    1588 allocs/op
PASS
ok  	github.com/maxmind/xgbshap	15.022s
```

</details>

<details>
<summary>after</summary>

```sh
$ go test -run=^$ -bench=. -benchmem -count=10 | tee new.txt
goos: darwin
goarch: arm64
pkg: github.com/maxmind/xgbshap
BenchmarkParseModel-10    	    1776	    676684 ns/op	  179090 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1773	    672871 ns/op	  179088 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1774	    674630 ns/op	  179088 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1774	    675294 ns/op	  179089 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1767	    674292 ns/op	  179089 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1767	    672589 ns/op	  179088 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1772	    674810 ns/op	  179088 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1770	    777569 ns/op	  179088 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1776	    672042 ns/op	  179088 B/op	    1115 allocs/op
BenchmarkParseModel-10    	    1770	    687123 ns/op	  179088 B/op	    1115 allocs/op
PASS
ok  	github.com/maxmind/xgbshap	13.984s
```

</details>